requirement.txt
appdirs==1.4.3
numpy==1.12.0
olefile==0.44
packaging==16.8
Pillow==4.0.0
pyparsing==2.2.0
scipy==0.19.0
six==1.10.0
tqdm==4.11.2
